adding the pep template to a place where folks can see it outside of the normal repo
PEPs get used in lots of different projects.  This will allow users,
customers, and the community to make suggestions/requests and allow
for guiding the roadmap.

Currently the "teams" are focused around our offerings:
- System (mmarineau, philips, polvi)
- Updates (robszumski, sym3tri)
- Orchestration (philips, bcwaldon)
- Online (polvi, robszumski)

These teams would be responsible for initial review & acceptance/
rejection of before going on to a more broad review.